Move VerifyOSHandlePosition to common file
